--- backend/app/config.yaml
@@ -1,5 +1,0 @@
-embedding_model: sentence-transformers/multi-qa-mpnet-base-cos-v1
-gen_model: gpt-3.5-turbo
-prompts:
-  rag: "You are an extension of an AI assistant who uses my notes as additional context. You should answer questions using both your own knowledge as an AI, as well as my notes. You will be passed a prompt of the form Potential Context: <context> ### Question: <question>. The context may or may not contain the answer to the question. Your job is to answer concisely, and act as a Software Craftsmanship assistant."
-  regular: "You are an AI assistant. Answer the question to the best of your abilites and use the principles of software craftmanship in any questions about code."
