+++ backend/app/routers/notes.py
@@ -1,7 +1,9 @@
 from fastapi import APIRouter
-from gptcotts.markdown_processor import convert_to_sections, convert_to_markdown, convert_to_chunks
+from gptcotts.indexing import update_notes
+from gptcotts.markdown_processor import (convert_to_chunks,
+                                         convert_to_markdown,
+                                         convert_to_sections)
 from gptcotts.s3_connector import get_object_from_s3, put_object_to_s3
-from gptcotts.pinecone_connector import update_notes
 from pydantic import BaseModel
 
 router = APIRouter(
@@ -25,15 +27,7 @@ def get_notes(request: NotesRequest):
     notes = get_object_from_s3("gptcotts-notes", f"{request.user_id}/{request.notes_class}.md")
     sections = convert_to_sections(notes)
 
-    return {
-        "sections": sections,
-        "metadata": {
-            "timing": {
-                "get_notes": get_object_from_s3.function_time, # type: ignore
-                "convert_to_sections": convert_to_sections.function_time # type: ignore
-            }
-        }
-    }
+    return {"sections": sections}
 
 @router.get("/refresh")
 def refresh_notes(request: UpdateNotesRequest):
@@ -42,17 +36,6 @@ def refresh_notes(request: UpdateNotesRequest):
     put_object_to_s3("gptcotts-notes", f"{request.user_id}/{request.notes_class}.md", markdown)
     notes_data = convert_to_chunks(markdown, request.notes_class)
     update_notes(request.pinecone_index, request.pinecone_namespace, notes_data)
-
-
-    return {
-            "metadata": {
-                "timing": {
-                    "convert_to_markdown": convert_to_markdown.function_time, # type: ignore
-                    "put_object_to_s3": put_object_to_s3.function_time, # type: ignore
-                    "convert_to_chunks": convert_to_chunks.function_time, # type: ignore
-                    "update_notes": update_notes.function_time # type: ignore
-                }
-            }
-    }
+    return {"status": "success"}
