--- devops/default.conf
@@ -1,16 +1,0 @@
-server {
-  listen 80;
-
-  server_name gptcotts.uk;
-
-  location /api/ {
-    rewrite ^/api/(.*)/$ $1 break;
-    proxy_pass http://localhost:8000/;
-  }
-
-
-  location / {
-    proxy_pass http://localhost:3000;
-  }
-}
-
