--- devops/default.template.conf
@@ -1,16 +1,0 @@
-server {
-  listen 80;
-
-  server_name {HOST};
-
-  location /api/ {
-    rewrite ^/api/(.*)/$ $1 break;
-    proxy_pass http://localhost:8000/;
-  }
-
-
-  location / {
-    proxy_pass http://localhost:3000;
-  }
-}
-
