+++ frontend/.env.production
@@ -1,1 +1,1 @@
-NEXT_PUBLIC_API_URL="/api"
+NEXT_PUBLIC_API_URL="https://nvksx3tqtj6g2qheiaeawgg3we0hklba.lambda-url.eu-north-1.on.aws/gptcotts/"
